UnitTestGeneration.Difficult.Tests.Gemini.Prompt1/SudokuSolverTests.cs
using UnitTestGeneration.Difficult.App;

namespace UnitTestGeneration.Difficult.Tests.Gemini.Prompt1;

public class SudokuSolverTests
{
    // Tests for FindEmpty
    [Fact]
    public void FindEmpty_FindsFirstEmptyCell()
    {
        int[][] testBoard = {
            new int[] { 0, 1, 0, 0, 0, 0, 0, 0, 0 }, // ... rest of your board
        };

        int[] expected = { 0, 0 }; // Coordinates of the first empty cell
        int[] result = SudokuSolver.FindEmpty(testBoard);

        Assert.Equal(expected, result);
    }

    [Fact]
    public void FindEmpty_NoEmptyCells_ReturnsNull()
    {
        int[][] testBoard = {
            new int[] { 1, 2, 3, 4, 5, 6, 7, 8, 9 }, // ... rest of your board (filled)
        };

        int[] result = SudokuSolver.FindEmpty(testBoard);
        Assert.Null(result);
    }

    // // Tests for IsPossible
    // [Fact]
    // public void IsPossible_ValidPlacement_ReturnsTrue()
    // {
    //     int[][] testBoard = { ... }; // A partially filled grid
    //     int[] targetCell = { 1, 2 }; // Coordinates of an empty cell
    //     int testNumber = 5; 
    //
    //     var result = SudokuSolver.IsPossible(testBoard, targetCell, testNumber);
    //     Assert.True(result);
    // }
    //
    // // ... More tests for IsPossible with invalid placements
    //
    // // Tests for SolveTheGrid (Integration-style)
    // [Fact]
    // public void SolveTheGrid_SolvesSimpleSudoku()
    // {
    //     int[][] testBoard = { ... }; // Simple, partially filled Sudoku
    //     int[][] expectedSolution = { ... }; // Corresponding solution
    //
    //     var result = SudokuSolver.SolveTheGrid(testBoard);
    //
    //     Assert.True(result);  // Assert it successfully solved
    //     Assert.Equal(expectedSolution, testBoard); // Compare with the expected solution
    // }
}